Separate tests in indiviual files
