+++ test/fieldValidators/test-multiSelectField.js
@@ -1,0 +1,9 @@
+var assert = require('assert');
+var expect = require('expect.js');
+
+var validators = require('../../lib/field_validators');
+var Schema = require('../../lib/schema');
+
+describe('Multi select field', function() {
+    
+});
